Created Updater class and added theme loader
Made a class for Updating, I haven't tested it but I'm fairly sure it should work.

Added a method to load the theme from settings, and implemented it in the GUI. [ALERT] This feature will not work until @avivhavivyan 's work on the themes folder and such features are added.

Added version file.

Co-Authored-By: avivhavivyan <60924491+avivhavivyan@users.noreply.github.com>